Move backfill code to `pkg/backfill` package
